+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [4.0.6] - 2025-07-10
 
+### Changed
+
+- Publish source files
+
 
 ## [4.0.5] - 2025-05-13